fix for removeUser

+++ wet.c
@@ -71,9 +71,21 @@ void* addUserMin(const char* name) {
 void* removeUser(const char* id) {
 	PGresult *res;
 	char cmd[200];
+	sprintf(cmd, "select id from users where id=%s",id);
+	res = PQexec(conn, cmd);
+	if (!res || PQresultStatus(res) != PGRES_TUPLES_OK) {
+		fprintf(stderr, "Error executing query: %s\n",
+				PQresultErrorMessage(res));
+		PQclear (res);
+		return;
+	}
+	if(PQntuples(res)==0){
+		printf("%s",ILL_PARAMS);
+		return;
+	}
 	sprintf(cmd,
-			"delete from users where id=%d;delete from photos where user_id=%d",
-			atoi(id), atoi(id));
+			"delete from users where id=%d;delete from photos where user_id=%d;delete from tags where user_id=%d",
+			atoi(id), atoi(id),atoi(id));
 	res = PQexec(conn, cmd);
 	if (!res || PQresultStatus(res) != PGRES_COMMAND_OK) {
 		fprintf(stderr, "Error executing query: %s\n",
@@ -243,7 +255,7 @@ void* mostCommonTags(const char* k) {
 }
 void* similarPhotos(const char* k, const char* j) {
 	PGresult *res;
-	char cmd[200];
+	char cmd[500];
 	sprintf(cmd,"select user_id,n1.name,photo_id from (select user_id,photo_id,count(user_id) from (select t1.user_id,t1.photo_id,count(t1.user_id) from tags as t1,tags as t2 where t1.info = t2.info and not(t1.user_id = t2.user_id and t1.photo_id = t2.photo_id) group by t1.user_id,t2.user_id,t1.photo_id,t2.photo_id) as tagtag where tagtag.count > %s group by user_id,photo_id having count(user_id)>0) as res,users as n1 where n1.id = res.user_id and count > %s order by user_id asc,photo_id desc",j,k);
 	res = PQexec(conn, cmd);
 	if (!res || PQresultStatus(res) != PGRES_TUPLES_OK) {
@@ -261,7 +273,7 @@ void* similarPhotos(const char* k, const char* j) {
 }
 void* autoPhotoOnTagOn() {
 	PGresult *res;
-	char cmd[400];
+	char cmd[200];
 	sprintf(cmd,"CREATE OR REPLACE FUNCTION insert_missing_photo() RETURNS TRIGGER AS $$ BEGIN IF(NEW.photo_id NOT IN (SELECT photo_id FROM Tags)) THEN INSERT INTO photos VALUES(NEW.photo_id,NEW.user_id); END IF; RETURN NEW; END; $$ LANGUAGE plpgsql; CREATE TRIGGER auto_photo BEFORE INSERT ON tags FOR EACH ROW EXECUTE PROCEDURE insert_missing_photo();");
 	res = PQexec(conn, cmd);
 	if (!res || PQresultStatus(res) != PGRES_COMMAND_OK) {
